requirement.txt
# Django
django==1.9.6
django-model-utils==2.5
django-compressor==2.0
django-extensions==1.6.7

phonenumbers==7.3.1
id-phonenumbers==0.1.4

bpython==0.15

django-thumbnails==0.1.3
django-redis==4.2.0
pillow==3.0

hashids==1.1.0

djangorestframework==3.3.3
python-dateutil==2.5.3
django-bower==5.1.0
libsass==0.11.1
unicodecsv==0.14.1

pytz==2016.4
django-redis-cache==1.6.5

rq==0.6.0
django-rq==0.9.1
python-gcm==0.4
